Extract dedicated tree components
